Add files via upload
This playbook contains a list of tasks to be executed by Ansible on a group of servers defined in the hosts field.

The become field specifies that the tasks should be executed with superuser privileges.

The tasks field contains a list of tasks to be executed. Each task has a name field that describes the task, and an action field that specifies the action to be taken.

The first task installs the Apache web server using the apt module. The apt module is used to manage packages on Debian-based systems. The name field specifies the name of the package to be installed (apache2), and the state field specifies that the package should be installed if it is not already present (present).

The second task starts the Apache web server using the service module. The service module is used to manage services on systems that use System V init or systemd. The name field specifies the name of the service (apache2), and the state field specifies that the service should be started (started).

The third task enables the Apache web server to start on boot using the systemd module. The systemd module is used to manage services on systems that use systemd. The name field specifies the name of the service (apache2), and the state field specifies that the service should be enabled to start on boot (enabled).

The fourth task creates a virtual host configuration file using a template. The template module is used to generate files from templates. The src field specifies the location of the template file (templates/virtualhost.conf.j2), and the dest field specifies the destination of the generated file (/etc/apache2/sites-available/example.com.conf).

The fifth task enables the SSL module for the Apache web server using the apache2_module module. The apache2_module module is used to manage Apache modules. The name field specifies the name of the module (ssl), and the state field specifies that the module should be enabled (enabled).

The sixth task restarts the Apache web server using the service module. The name field specifies the name of the service (apache2), and the state field specifies that the service should be restarted (restarted). This is done to apply the changes made by the previous tasks.